obdemo/requirements.txt
django-static==1.4.0
slimmer==0.1.30
